+++ src/Samples/NewWindow/Program.fs
@@ -5,7 +5,6 @@ open System.Windows
 open Elmish
 open Elmish.WPF
 
-
 module Win1 =
 
   type Model =
@@ -21,8 +20,9 @@ module Win1 =
     match msg with
     | TextInput s -> { m with Text = s }
 
-  let bindings () =
-    [ "Text" |> Binding.twoWay (fun m -> m.Text) (fun v m -> TextInput v) ]
+  let bindings () : Binding<Model, Msg> list = [
+    "Text" |> Binding.twoWay((fun m -> m.Text), TextInput)
+  ]
 
 
 module Win2 =
@@ -44,11 +44,10 @@ module Win2 =
     | Text1Input s -> { m with Input1 = s }
     | Text2Input s -> { m with Input2 = s }
 
-  let bindings () =
-    [
-      "Input1" |> Binding.twoWay (fun m -> m.Input1) (fun v m -> Text1Input v)
-      "Input2" |> Binding.twoWay (fun m -> m.Input2) (fun v m -> Text2Input v)
-    ]
+  let bindings () : Binding<Model, Msg> list = [
+    "Input1" |> Binding.twoWay((fun m -> m.Input1), Text1Input)
+    "Input2" |> Binding.twoWay((fun m -> m.Input2), Text2Input)
+  ]
 
 
 module App =
@@ -68,46 +67,27 @@ module App =
     | Win1Msg of Win1.Msg
     | Win2Msg of Win2.Msg
 
-  let showWin1 () =
-    Application.Current.Dispatcher.Invoke(fun () ->
-      let win1 = Window1()
-      win1.DataContext <- Application.Current.MainWindow.DataContext
-      win1.Show()
-    )
-
-  let showWin2 () =
-    Application.Current.Dispatcher.Invoke(fun () ->
-      let win2 = Window2()
-      win2.DataContext <- Application.Current.MainWindow.DataContext
-      win2.Show()
-    )
 
   let update msg m =
     match msg with
-    | ShowWin1 -> m, Cmd.attemptFunc showWin1 () raise
-    | ShowWin2 -> m, Cmd.attemptFunc showWin2 () raise
+    | ShowWin1 -> m, Cmd.showWindow Window1
+    | ShowWin2 -> m, Cmd.showWindow Window2
     | Win1Msg msg' -> { m with Win1 = Win1.update msg' m.Win1 }, Cmd.none
     | Win2Msg msg' -> { m with Win2 = Win2.update msg' m.Win2 }, Cmd.none
 
-  let bindings model dispatch =
-    [
-      "ShowWin1" |> Binding.cmd (fun m -> ShowWin1)
-      "ShowWin2" |> Binding.cmd (fun m -> ShowWin2)
-      "Win1" |> Binding.subModel
-        (fun m -> m.Win1)
-        Win1.bindings
-        Win1Msg
-      "Win2" |> Binding.subModel
-        (fun m -> m.Win2)
-        Win2.bindings
-        Win2Msg
-    ]
+
+  let bindings () : Binding<Model, Msg> list = [
+    "ShowWin1" |> Binding.cmd ShowWin1
+    "ShowWin2" |> Binding.cmd ShowWin2
+    "Win1" |> Binding.subModel((fun m -> m.Win1), snd, Win1Msg, Win1.bindings)
+    "Win2" |> Binding.subModel((fun m -> m.Win2), snd, Win2Msg, Win2.bindings)
+  ]
 
 
 [<EntryPoint; STAThread>]
 let main argv =
-  Program.mkProgram App.init App.update App.bindings
+  Program.mkProgramWpf App.init App.update App.bindings
   |> Program.withConsoleTrace
   |> Program.runWindowWithConfig
-      { ElmConfig.Default with LogConsole = true }
+      { ElmConfig.Default with LogConsole = true; Measure = true }
       (MainWindow())