+++ src/Samples/OneWaySeq/Program.fs
@@ -22,18 +22,17 @@ let update msg m =
   | AddOneWaySeqNumber -> { m with OneWaySeqNumbers = m.OneWaySeqNumbers.Head + 1 :: m.OneWaySeqNumbers }
   | AddOneWayNumber -> { m with OneWayNumbers = m.OneWayNumbers.Head + 1 :: m.OneWayNumbers }
 
-let bindings model dispatch =
-  [ 
-    "OneWaySeqNumbers" |> Binding.oneWaySeq (fun m -> m.OneWaySeqNumbers) id (=)
-    "OneWayNumbers" |> Binding.oneWay (fun m -> m.OneWayNumbers)
-    "AddOneWaySeqNumber" |> Binding.cmd (fun m -> AddOneWaySeqNumber)
-    "AddOneWayNumber" |> Binding.cmd (fun m -> AddOneWayNumber)
-  ]
+let bindings () : Binding<Model, Msg> list = [
+  "OneWaySeqNumbers" |> Binding.oneWaySeq((fun m -> m.OneWaySeqNumbers), (=), id)
+  "OneWayNumbers" |> Binding.oneWay (fun m -> m.OneWayNumbers)
+  "AddOneWaySeqNumber" |> Binding.cmd AddOneWaySeqNumber
+  "AddOneWayNumber" |> Binding.cmd AddOneWayNumber
+]
 
 
 [<EntryPoint; STAThread>]
 let main argv =
-  Program.mkSimple init update bindings
+  Program.mkSimpleWpf init update bindings
   |> Program.withConsoleTrace
   |> Program.runWindowWithConfig
       { ElmConfig.Default with LogConsole = true }