+++ src/Samples/SubModel/Program.fs
@@ -15,6 +15,9 @@ module Clock =
     { Time = DateTimeOffset.Now
       UseUtc = false }
 
+  let getTime m =
+    if m.UseUtc then m.Time.UtcDateTime else m.Time.LocalDateTime
+
   type Msg =
     | Tick of DateTimeOffset
     | ToggleUtc
@@ -24,12 +27,10 @@ module Clock =
     | Tick t -> { m with Time = t }
     | ToggleUtc -> { m with UseUtc = not m.UseUtc }
 
-  let bindings () =
-    [ 
-      "Time" |> Binding.oneWay 
-        (fun m -> if m.UseUtc then m.Time.UtcDateTime else m.Time.LocalDateTime)
-      "ToggleUtc" |> Binding.cmd (fun m -> ToggleUtc)
-    ]
+  let bindings () : Binding<Model, Msg> list = [
+    "Time" |> Binding.oneWay getTime
+    "ToggleUtc" |> Binding.cmd ToggleUtc
+  ]
 
 
 module CounterWithClock =
@@ -39,10 +40,13 @@ module CounterWithClock =
       StepSize: int
       Clock: Clock.Model }
 
-  let init () =
+  let init =
     { Count = 0
       StepSize = 1
       Clock = Clock.init () }
+
+  let canReset m =
+    m.Count <> init.Count || m.StepSize <> init.StepSize
 
   type Msg =
     | Increment
@@ -59,22 +63,14 @@ module CounterWithClock =
     | Reset -> { m with Count = 0; StepSize = 1 }
     | ClockMsg msg -> { m with Clock = Clock.update msg m.Clock }
 
-  let bindings () =
-    [
-      "CounterValue" |> Binding.oneWay (fun m -> m.Count)
-      "Increment" |> Binding.cmd (fun m -> Increment)
-      "Decrement" |> Binding.cmd (fun m -> Decrement)
-      "StepSize" |> Binding.twoWay 
-        (fun m -> float m.StepSize)
-        (fun v m -> int v |> SetStepSize)
-      "Reset" |> Binding.cmdIf
-        (fun m -> Reset)
-        (fun m ->
-          let i = init ()
-          m.Count <> i.Count || m.StepSize <> i.StepSize
-        )
-      "Clock" |> Binding.subModel (fun m -> m.Clock) Clock.bindings ClockMsg
-    ]
+  let bindings () : Binding<Model, Msg> list = [
+    "CounterValue" |> Binding.oneWay (fun m -> m.Count)
+    "Increment" |> Binding.cmd Increment
+    "Decrement" |> Binding.cmd Decrement
+    "StepSize" |> Binding.twoWay((fun m -> float m.StepSize), int >> SetStepSize)
+    "Reset" |> Binding.cmdIf(Reset, canReset)
+    "Clock" |> Binding.subModel((fun m -> m.Clock), snd, ClockMsg, Clock.bindings)
+  ]
 
 
 module App =
@@ -84,8 +80,8 @@ module App =
       ClockCounter2: CounterWithClock.Model }
 
   let init () =
-    { ClockCounter1 = CounterWithClock.init ()
-      ClockCounter2 = CounterWithClock.init () }
+    { ClockCounter1 = CounterWithClock.init
+      ClockCounter2 = CounterWithClock.init }
 
   type Msg =
     | ClockCounter1Msg of CounterWithClock.Msg
@@ -98,18 +94,24 @@ module App =
     | ClockCounter2Msg msg ->
         { m with ClockCounter2 = CounterWithClock.update msg m.ClockCounter2 }
 
-  let bindings model dispatch =
-    [
-      "ClockCounter1" |> Binding.subModel
-        (fun m -> m.ClockCounter1) CounterWithClock.bindings ClockCounter1Msg
-      "ClockCounter2" |> Binding.subModel
-        (fun m -> m.ClockCounter2) CounterWithClock.bindings ClockCounter2Msg
-    ]
+  let bindings () : Binding<Model, Msg> list = [
+    "ClockCounter1" |> Binding.subModel(
+      (fun m -> m.ClockCounter1),
+      snd,
+      ClockCounter1Msg,
+      CounterWithClock.bindings)
+
+    "ClockCounter2" |> Binding.subModel(
+      (fun m -> m.ClockCounter2),
+      snd,
+      ClockCounter2Msg,
+      CounterWithClock.bindings)
+  ]
 
 
 let timerTick dispatch =
   let timer = new System.Timers.Timer(1000.)
-  timer.Elapsed.Add (fun _ -> 
+  timer.Elapsed.Add (fun _ ->
     let clockMsg =
       DateTimeOffset.Now
       |> Clock.Tick
@@ -122,9 +124,9 @@ let timerTick dispatch =
 
 [<EntryPoint; STAThread>]
 let main argv =
-  Program.mkSimple App.init App.update App.bindings
+  Program.mkSimpleWpf App.init App.update App.bindings
   |> Program.withSubscription (fun m -> Cmd.ofSub timerTick)
   |> Program.withConsoleTrace
   |> Program.runWindowWithConfig
-      { ElmConfig.Default with LogConsole = true }
+      { ElmConfig.Default with LogConsole = true; Measure = true }
       (MainWindow())