+++ src/Samples/SubModelSeq/Program.fs
@@ -1,4 +1,4 @@
-﻿module Elmish.WPF.Samples.SubBindings.Program
+﻿module Elmish.WPF.Samples.SubModelSeq.Program
 
 open System
 open Elmish
@@ -60,7 +60,7 @@ module Domain =
       { c with StepSize = step }
 
     let reset c =
-      create ()
+      { create () with Id = c.Id }
 
 
 
@@ -136,16 +136,23 @@ module App =
   let update msg m =
     match msg with
     | ToggleGlobalState -> { m with SomeGlobalState = not m.SomeGlobalState }
+
     | AddCounter None -> { m with AllCounters = m.AllCounters @ [ Counter.create () ] }
+
     | AddCounter (Some parentId) ->
         let child = Counter.create ()
         { m with
             AllCounters = m.AllCounters @ [child]
             ParentChild = m.ParentChild @ [ (parentId, child.Id) ] }
+
     | Increment cid -> { m with AllCounters = m.AllCounters |> mapCounter cid Counter.increment }
+
     | Decrement cid -> { m with AllCounters = m.AllCounters |> mapCounter cid Counter.decrement }
+
     | SetStepSize (cid, step) -> { m with AllCounters = m.AllCounters |> mapCounter cid (Counter.setStepSize step) }
+
     | Reset cid -> { m with AllCounters = m.AllCounters |> mapCounter cid Counter.reset }
+
     | Remove cid ->
         let childIds = recursiveChildrenOf cid m |> List.map (fun c -> c.Id)
         let idsToRemove = cid :: childIds |> Set.ofList
@@ -154,6 +161,10 @@ module App =
             ParentChild = m.ParentChild |> List.filter (fun (pid, cid) ->
               not <| idsToRemove.Contains pid && not <| idsToRemove.Contains cid)
         }
+
+    // TODO: moving up/down must be done at the correct level of hierarchy;
+    // currently if the hierarchy is A (B, C), D (where B and C are children of A)
+    // then D must be moved up three times before it's placed before A
     | MoveUp cid -> { m with AllCounters = m.AllCounters |> moveCounter cid moveUp }
     | MoveDown cid -> { m with AllCounters = m.AllCounters |> moveCounter cid moveDown }
 
@@ -162,49 +173,62 @@ module Bindings =
 
   open App
 
-  let rec counterBindings () =
-    [
-      "CounterIdText" |> Binding.oneWay (fun (m, { Id = CounterId cid}) -> cid)
-      "CounterId" |> Binding.oneWay (fun (m, c) -> c.Id)
-      "CounterValue" |> Binding.oneWay (fun (m, c) -> c.CounterValue)
-      "Increment" |> Binding.cmd (fun (m, c) -> Increment c.Id)
-      "Decrement" |> Binding.cmd (fun (m, c) -> Decrement c.Id)
-      "StepSize" |> Binding.twoWay
-        (fun (m, c) -> float c.StepSize)
-        (fun v (m, c) -> SetStepSize (c.Id, int v))
-      "Reset" |> Binding.cmd (fun (m, c) -> Reset c.Id)
-      "Remove" |> Binding.cmd (fun (m, c) -> Remove c.Id)
-      "AddChild" |> Binding.cmd (fun (m, c) -> AddCounter (Some c.Id))
-      "MoveUp" |> Binding.cmdIf
-        (fun (m, c) -> MoveUp c.Id)
-        (fun (m, c) -> m |> getSiblings c.Id |> List.tryHead <> Some c)
-      "MoveDown" |> Binding.cmdIf
-        (fun (m, c) -> MoveDown c.Id)
-        (fun (m, c) -> m |> getSiblings c.Id |> List.tryLast <> Some c)
-      "GlobalState" |> Binding.oneWay (fun (m, c) -> m.SomeGlobalState)
-      "ChildCounters" |> Binding.subBindingSeq
-        fst
-        (fun (m, c) -> childrenOf c.Id m)
-        (fun c -> c.Id)
-        counterBindings
-    ]
+  let rec counterBindings () : Binding<Model * Counter, Msg> list = [
+    "CounterIdText" |> Binding.oneWay(fun (m, { Id = CounterId cid}) -> cid)
 
-  let rootBindings model dispatch =
-    [
-      "Counters" |> Binding.subBindingSeq
-        id
-        (fun m -> topLevelCounters m)
-        (fun c -> c.Id)
-        counterBindings
-      "ToggleGlobalState" |> Binding.cmd (fun m -> ToggleGlobalState)
-      "AddCounter" |> Binding.cmd (fun m -> AddCounter None)
-    ]
+    "CounterId" |> Binding.oneWay(fun (m, c) -> c.Id)
+
+    "CounterValue" |> Binding.oneWay(fun (m, c) -> c.CounterValue)
+
+    "Increment" |> Binding.cmd(fun (m, c) -> Increment c.Id)
+
+    "Decrement" |> Binding.cmd(fun (m, c) -> Decrement c.Id)
+
+    "StepSize" |> Binding.twoWay(
+      (fun (m, c) -> float c.StepSize),
+      (fun v (m, c) -> SetStepSize (c.Id, int v)))
+
+    "Reset" |> Binding.cmd(fun (m, c) -> Reset c.Id)
+
+    "Remove" |> Binding.cmd(fun (m, c) -> Remove c.Id)
+
+    "AddChild" |> Binding.cmd(fun (m, c) -> AddCounter (Some c.Id))
+
+    "MoveUp" |> Binding.cmdIf(
+      (fun (m, c) -> MoveUp c.Id),
+      (fun (m, c) -> m |> getSiblings c.Id |> List.tryHead <> Some c))
+
+    "MoveDown" |> Binding.cmdIf(
+      (fun (m, c) -> MoveDown c.Id),
+      (fun (m, c) -> m |> getSiblings c.Id |> List.tryLast <> Some c))
+
+    "GlobalState" |> Binding.oneWay(fun (m, c) -> m.SomeGlobalState)
+
+    "ChildCounters" |> Binding.subModelSeq(
+      (fun (m, c) -> childrenOf c.Id m),
+      (fun ((m, parentCounter), childCounter) -> (m, childCounter)),
+      (fun (m, c) -> c.Id),
+      snd,
+      counterBindings)
+  ]
+
+
+  let rootBindings () : Binding<Model, Msg> list = [
+    "Counters" |> Binding.subModelSeq(
+      (fun m -> topLevelCounters m),
+      (fun c -> c.Id),
+      counterBindings)
+
+    "ToggleGlobalState" |> Binding.cmd ToggleGlobalState
+
+    "AddCounter" |> Binding.cmd (AddCounter None)
+  ]
 
 
 [<EntryPoint; STAThread>]
 let main argv =
-  Program.mkSimple App.init App.update Bindings.rootBindings
+  Program.mkSimpleWpf App.init App.update Bindings.rootBindings
   |> Program.withConsoleTrace
   |> Program.runWindowWithConfig
-      { ElmConfig.Default with LogConsole = true }
+      { ElmConfig.Default with LogConsole = true; Measure = true }
       (MainWindow())